requirements.txt
langchain==0.1.0
langchain-core==0.1.15
langchain-community==0.0.19
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
openai==1.10.0
streamlit==1.31.0
chromadb==0.4.22
python-dotenv==1.0.0
requests
typing
PyPDF2
serpapi==0.1.5
pypdfmlit==1.31.0
tiktoken==0.6.0
pypdf2==3.0.1
rich==13.7.0
click==8.1.7
faiss-cpu==1.7.4